move definition of prepareModifyReplicationGroupInput
